requirements.txt
torch
numpy
scipy
pyro-ppl
tqdm
matplotlib
netCDF4
pandas
xarray[io]
xarray
click==8.0.4
black==22.1.0
nose2
pylint==2.15.5
functorch
